// frontend/src/api/api.ts
export interface LoginDto {
    /**
     * The email address of the user
     * @type {string}
     * @memberof LoginDto
     */
    'email': string;
    /**
     * The password of the user
     * @type {string}
     * @memberof LoginDto
     */
    'password': string;
}
/**
 * 
 * @export
 * @interface LoginResponseDto
 */
export interface LoginResponseDto {
    /**
     * 
     * @type {User}
     * @memberof LoginResponseDto
     */
    'user': User;
    /**
     * 
     * @type {string}
     * @memberof LoginResponseDto
     */
    'accessToken': string;
}
/**
 * 
 * @export
 * @interface SignUpDto
 */
export interface SignUpDto {
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'phone': string;
    /**
     * 
     * @type {number}
     * @memberof SignUpDto
     */
    'age': number;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    'password': string;
}
// ...
export class AuthApi extends BaseAPI {
    /**
     * 
     * @summary Login to your account
     * @param {LoginDto} loginDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authControllerLogin(loginDto: LoginDto, options?: RawAxiosRequestConfig) {
        return AuthApiFp(this.configuration).authControllerLogin(loginDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Logout from your account
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authControllerLogout(options?: RawAxiosRequestConfig) {
        return AuthApiFp(this.configuration).authControllerLogout(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Register a user and trigger OTP email
     * @param {SignUpDto} signUpDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authControllerRegisterUser(signUpDto: SignUpDto, options?: RawAxiosRequestConfig) {
        return AuthApiFp(this.configuration).authControllerRegisterUser(signUpDto, options).then((request) => request(this.axios, this.basePath));
    }
}